Split code into more scripts for sanity
