+++ .pylintrc
@@ -1,7 +1,7 @@
 [MASTER]
 # Add the directory containing merge_tool.py to the Python path
-init-hook='import sys, os; sys.path.insert(0, os.path.abspath("."))'
+init-hook='import sys, os; sys.path.insert(0, os.path.abspath("scripts"))'
 
 [MESSAGES CONTROL]
 # Disable the message about missing module docstring
-disable=C0114,C0415,R0913,R0917,C0301,R0914,R0911,R0912,R0915,W0613,W1203,C0302,R0904
+disable=C0114,C0415,R0913,R0917,C0301,R0914,R0911,R0912,R0915,W0613,W1203,C0302,R0904,C0103
